ISSUE-525/764: Local @topics variable does not contain all the subscription information
The leader consumer thread needs to have knowledge about
all siblings consumer subscriptions in order to be perform
the correct partition assignments.
For consumers with different subscription this is not possible
through the local `@topics`. A way to solve this is by obtaining
this info from the cluster itself through the `@members` var.